Divided code in modules
Implemented MatchStarted command
Divided code into modules
